Reduced sonar code smells.

+++ src/main/java/growthcraft/milk/item/MilkingBucketItem.java
@@ -162,7 +162,7 @@ public class MilkingBucketItem extends Item implements DispensibleContainerItem 
                 p_150717_.playSound(p_150716_, p_150718_, SoundEvents.FIRE_EXTINGUISH, SoundSource.BLOCKS, 0.5F, 2.6F + (p_150717_.random.nextFloat() - p_150717_.random.nextFloat()) * 0.8F);
 
                 for(int l = 0; l < 8; ++l) {
-                    p_150717_.addParticle(ParticleTypes.LARGE_SMOKE, (double)i + random.nextDouble(), (double)j + random.nextDouble(), (double)k + random.nextDouble(), 0.0D, 0.0D, 0.0D);
+                    p_150717_.addParticle(ParticleTypes.LARGE_SMOKE, i + random.nextDouble(), j + random.nextDouble(), k + random.nextDouble(), 0.0D, 0.0D, 0.0D);
                 }
 
                 return true;
